Fixes issue 2963, ref socks_connecter.cpp:158

+++ src/socks_connecter.cpp
@@ -158,7 +158,7 @@ void zmq::socks_connecter_t::in_event ()
                 //  Attach the engine to the corresponding session object.
                 send_attach (session, engine);
 
-                socket->event_connected (endpoint, (int) s);
+                socket->event_connected (endpoint, s);
 
                 rm_fd (handle);
                 s = -1;
